엔티티 pk 타입 변경(int->long), 엔티티chkstatus enum 타입으로 변경

+++ src/main/java/com/thxpapa/juneberrydiary/domain/blog/BlogUser.java
@@ -15,7 +15,7 @@ public class BlogUser extends BaseEntity {
     @Id
     @Column(name="blog_user_uid")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int blogUserUid;
+    private Long blogUserUid;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "blogId")
