+++ src/main/java/com/thxpapa/juneberrydiary/repository/blogRepository/BlogUserRepository.java
@@ -3,5 +3,5 @@ package com.thxpapa.juneberrydiary.repository.blogRepository;
 import com.thxpapa.juneberrydiary.domain.blog.BlogUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogUserRepository extends JpaRepository<BlogUser, Integer> {
+public interface BlogUserRepository extends JpaRepository<BlogUser, Long> {
 }
