+++ src/main/java/com/thxpapa/juneberrydiary/repository/geoRepository/AlbumRepository.java
@@ -3,5 +3,5 @@ package com.thxpapa.juneberrydiary.repository.geoRepository;
 import com.thxpapa.juneberrydiary.domain.geo.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
+public interface AlbumRepository extends JpaRepository<Album, Long> {
 }
